mod fsm in separate file
