+++ src/nbt/mod.rs
@@ -1,193 +1,11 @@
 use byteorder::{BigEndian, ReadBytesExt};
 use core::{panic, str};
-use std::io::{Cursor, Seek, SeekFrom};
+use std::io::Cursor;
 
 use thiserror::Error;
 use std::io;
 
-mod fsm {
-    use crate::nbt;
-    use std::io::{Cursor, Seek, SeekFrom};
-
-    pub enum ParseNbtFsm {
-        Normal,
-        NoAction,
-        List,
-        EndOfFile
-    }
-    
-    
-    struct NbtListParser {
-        list_tag_id: nbt::NbtTagId,
-        list_len: i32,
-        list_elem_count: i32,
-    }
-    
-    impl NbtListParser {
-        pub fn new() -> NbtListParser {
-            NbtListParser { list_tag_id: nbt::NbtTagId::End, list_len: 0, list_elem_count: 0 }
-        }
-    
-        pub fn set_id(&mut self, tag_id: nbt::NbtTagId) {
-            self.list_tag_id = tag_id;
-        }
-    
-        pub fn tag_id(&self) -> &nbt::NbtTagId {
-            &self.list_tag_id
-        }
-    
-        pub fn set_len(&mut self, len: i32) {
-            self.list_len = len;
-        }
-    
-        pub fn increment(&mut self) {
-            self.list_elem_count = self.list_elem_count + 1;
-        }
-    
-        pub fn reset(&mut self) {
-            self.list_tag_id = nbt::NbtTagId::End;
-            self.list_len = 0;
-            self.list_elem_count = 0;
-        }
-    
-        pub fn is_end(&self) -> bool {
-            self.list_elem_count >= self.list_len - 1
-        }
-    }
-    
-    pub struct NbtParser {
-        state: ParseNbtFsm,
-        list_parser: NbtListParser,
-        cursor: Cursor<Vec<u8>>,
-        index: usize,
-        tree_depth: i64,
-    }
-    
-    impl NbtParser {
-        pub fn new(state: ParseNbtFsm, cursor: Cursor<Vec<u8>>) -> NbtParser {
-            NbtParser { state: state, 
-                        list_parser: NbtListParser::new(),
-                        cursor: cursor,
-                        index: 0,
-                        tree_depth: 0
-                     }
-        }
-    
-        pub fn change_state_to(&mut self, state: ParseNbtFsm) {
-            self.state = state;
-        }
-    
-        pub fn state(&self) -> &ParseNbtFsm {
-            &self.state
-        }
-    
-        pub fn cursor(&mut self) -> Cursor<Vec<u8>> {
-            self.cursor.clone() //TODO: use &mut self.cursor
-        }
-    
-        pub fn index(&self) -> &usize {
-            &self.index
-        }
-    
-        pub fn increment_index(&mut self) {
-            self.index = self.index + 1;
-        }
-    }
-    
-    pub fn fsm_parse(test_sequence : &mut nbt::NbtTagSequence, nbt_parser: &mut NbtParser) {//-> (NbtTagId, String, NbtTagType) {
-        
-        let mut cursor = nbt_parser.cursor();
-        let mut tag_id;
-        let mut depth_delta= 0;
-        let total_bytes = cursor.seek(SeekFrom::End(0)).unwrap();
-        
-        cursor.seek(SeekFrom::Start(0)).unwrap();
-        loop {
-
-            let byte_start = cursor.position();   
-            let mut tag_name = String::new();
-            let mut tag_value = nbt::NbtTagType::End(None);
-            
-            nbt_parser.tree_depth+=depth_delta;
-            depth_delta = 0;
-
-            match nbt_parser.state() {
-                ParseNbtFsm::Normal => {
-                    tag_id = nbt::NbtTag::parse_nbt_tag_id(&mut cursor).unwrap();
-
-                    if let nbt::NbtTagId::End = tag_id {
-                        depth_delta -= 1;
-                    }
-                    else {
-                        tag_name = nbt::NbtTag::parse_nbt_tag_string(&mut cursor).unwrap();    
-                        tag_value = nbt::NbtTag::parse_nbt_tag(&mut cursor, &tag_id).unwrap();
-
-                        if let nbt::NbtTagType::List(ref list_elem_tag_ids) = tag_value {
-                            nbt_parser.list_parser.set_id(list_elem_tag_ids.0);
-                            nbt_parser.list_parser.set_len(list_elem_tag_ids.1);
-                            nbt_parser.change_state_to(ParseNbtFsm::List); 
-                            depth_delta += 1;
-                        }
-
-                        if let nbt::NbtTagId::Compound = tag_id {
-                            depth_delta += 1;
-                        }
-                    }
-                    
-                },
-
-                ParseNbtFsm::List => {
-                    if nbt_parser.list_parser.is_end() {
-                        tag_id = *nbt_parser.list_parser.tag_id();
-                        nbt_parser.list_parser.reset();
-                        nbt_parser.change_state_to(ParseNbtFsm::Normal); 
-                        depth_delta -= 1;
-                    }
-                    else {
-                        nbt_parser.list_parser.increment(); 
-                        tag_id = *nbt_parser.list_parser.tag_id();            
-                    }
-
-                    tag_name = "".to_string();
-                    tag_value = nbt::NbtTag::parse_nbt_tag(&mut cursor, &tag_id).unwrap();
-
-                    if let nbt::NbtTagId::Compound = tag_id {
-                        depth_delta += 1;
-                    }
-                },
-
-                ParseNbtFsm::NoAction => {
-                    //shall never be reached. delete the state?
-                },
-
-                ParseNbtFsm::EndOfFile => {
-                    break;
-                },
-            }
-
-            let byte_end = cursor.position();
-
-
-
-            test_sequence.tags.push(nbt::NbtTag { name: tag_name, 
-                                            value: tag_value, 
-                                            byte_start: byte_start, 
-                                            byte_end: byte_end,
-                                            index: *nbt_parser.index(),
-                                            depth: nbt_parser.tree_depth});
-
-            nbt_parser.increment_index();
-
-            if byte_end >= total_bytes {
-                nbt_parser.change_state_to(ParseNbtFsm::EndOfFile);
-                break; //TODO Remove
-            }
-        }
-    }
-
-
-}
-
+mod fsm;
 
 #[derive(Error, Debug)]
 pub enum NbtReadError {
@@ -317,141 +135,12 @@ impl NbtTag {
     }
 
     pub fn parse_from_buf(cursor: &mut Cursor<Vec<u8>>) -> Result<NbtTagSequence, NbtReadError> {
-    
-/*         let (nbt_root_present, 
-            tag_id, 
-            tag_name) = NbtTag::nbt_root_present(cursor);
-
-        if nbt_root_present == false {
-            return NbtTag { name: tag_name, value: NbtTagType::End(None) };
-        }
-        else {
-            let mut compound_values = Vec::<NbtTag>::new();
-            let compound = NbtTag::traverse_nbt_tree(cursor);
-
-            compound_values.push(compound);
-            NbtTag { name: tag_name, value: NbtTagType::Compound(compound_values) }
-        } */
-
-        /* let nbt_tag = match NbtTag::traverse_nbt_tree(cursor, &mut nbt_index) {
-            Ok(nbt_tag) => nbt_tag,
-            Err(e) => return Err(e),
-        };      
- */     let mut nbt_parser = fsm::NbtParser::new(fsm::ParseNbtFsm::Normal, cursor.clone());
+        let mut nbt_parser = fsm::NbtParser::new(fsm::ParseNbtFsm::Normal, cursor.clone());
         let mut test_sequence = NbtTagSequence::new();
-        fsm::fsm_parse(&mut test_sequence, &mut nbt_parser);
+        fsm::parse(&mut test_sequence, &mut nbt_parser);
         
         Ok(test_sequence)
     }
-
-
-
-    
-   /*  fn traverse_nbt_tree(cursor: &mut Cursor<Vec<u8>>, test_sequence : &mut NbtTagSequence, nbt_parser: &mut NbtParser, nbt_index: &mut usize) { //-> Result<NbtReadError> {
-        
-        let mut finished_reading = false;
-
-        //let mut list_tag_id = NbtTagId::End;
-        //let mut list_len = 0;
-        //let mut list_elem_count = 0;
-        //let mut reading_list = false;
-
-        //nbt_parser.list_parser.reset();
-
-        let mut tag_id;
-
-        let total_bytes = cursor.seek(SeekFrom::End(0)).unwrap();
-        cursor.seek(SeekFrom::Start(0)).unwrap();
-
-        
-        
-        while finished_reading == false {
-
-            let byte_start = cursor.position();
-            
-            let mut tag_name = String::new();
-            let mut tag_value = NbtTagType::End(None);
-
-            if let ParseNbtFsm::List = nbt_parser.state() {
-                //a list is a sequence of NBT Tags without names and tags id.
-                
-                if nbt_parser.list_parser.is_end() {
-                    tag_id = *nbt_parser.list_parser.tag_id();
-                    nbt_parser.list_parser.reset();
-                    nbt_parser.change_state_to(ParseNbtFsm::Normal); 
-                }
-                else {
-                    nbt_parser.list_parser.increment(); 
-                    tag_id = *nbt_parser.list_parser.tag_id();            
-                }
-
-                tag_name = "".to_string();
-            }
-            else {
-                tag_id = NbtTag::parse_nbt_tag_id(cursor).unwrap();
-                if let NbtTagId::End = tag_id {
-                    // nothing to do
-                }
-                else {
-                    tag_name = NbtTag::parse_nbt_tag_string(cursor).unwrap();
-                }
-            }
-
-            if let NbtTagId::End = tag_id {
-                // nothing to do
-            }
-            else {
-                tag_value = NbtTag::parse_nbt_tag(cursor, &tag_id).unwrap();
-
-                if let NbtTagType::List(ref list_elem_tag_ids) = tag_value {
-                    nbt_parser.list_parser.set_id(list_elem_tag_ids.0);
-                    nbt_parser.list_parser.set_len(list_elem_tag_ids.1);
-                    nbt_parser.change_state_to(ParseNbtFsm::List); 
-                }
-            }
-            println!("-------------------------------");
-            println!("Tag ID: {:?}", tag_id);
-            println!("Tag Name: {:?}", tag_name);
-            println!("Tag Value: {:?}", tag_value); 
-
-
-            let byte_end = cursor.position();
-
-            let current_index = *nbt_index;
-            *nbt_index = current_index + 1;
-
-            test_sequence.tags.push(NbtTag { name: tag_name, 
-                                            value: tag_value, 
-                                            byte_start: byte_start, 
-                                            byte_end: byte_end,
-                                            index: current_index });
-
-            if byte_end >= total_bytes {
-                finished_reading = true;
-            }
-        }
-            
-    }
-
-     */
-
-/*     fn nbt_root_present(cursor: &mut Cursor<Vec<u8>>) -> (bool, Option<NbtTagId>, String) {
-        
-        let tag_id = match NbtTag::parse_nbt_tag_id(cursor) {
-            None => return (false, None, "".to_string()),
-            Some(tag_id) => tag_id,
-        };
-
-        let tag_name = match NbtTag::parse_nbt_tag_string(cursor) {
-            Ok(tag_name) => tag_name,
-            Err(e) => return (false, None, e),
-        };
-
-        match tag_id {
-            NbtTagId::Compound => (true, Some(tag_id), tag_name),
-            _ => (false, Some(tag_id), tag_name),
-        }
-    } */
 
     fn parse_nbt_tag_id(cursor: &mut Cursor<Vec<u8>>) -> Option<NbtTagId> {
         let id = cursor.read_u8().expect("Error reading byte from cursor");
